New Editor: Improve wording, especially in preferences
* Refer to new editor as "modern editor", and fix several places where
  it was still called the "experimental new editor".
* Refer to old editor as "legacy editor".
* Remove link to "learn more about the New Editor", as the aka has not
  been maintained and there is no useful documentation page right now.
* Change the "Open C# files in the New Editor" preference to read
  "Use the legacy text editor where available (not recommended)", and
  invert its effect. Move this preference to the end of its section.

In an upcoming commit, this will be the only way to enable the legacy
editor in preferences, and it will only work for file types that
explicitly maintain legacy editor support.

Users will continue to be able to use the "Open With" menu to pick the
editor they prefer at that time.